Fix Bug with dead alley branches in path list
Before this commit the path List contains dead alley branches. To fix
this problem all nodes have now a parent node. After calculating the
path, the finish node is taken and the result path is created by collecting
the parent node (of each node) till the root node. With this fix the dead alley
branches are removed.